+++ packages/apps-routing/src/fellowship.ts
@@ -4,7 +4,7 @@
 import type { TFunction } from 'i18next';
 import type { Route } from './types';
 
-import Component from '@polkadot/app-fellowship';
+import Component, { useCounter } from '@polkadot/app-fellowship';
 
 export default function create (t: TFunction): Route {
   return {
@@ -19,6 +19,7 @@ export default function create (t: TFunction): Route {
     group: 'governance',
     icon: 'people-arrows',
     name: 'fellowship',
-    text: t('nav.fellowship', 'Fellowship', { ns: 'apps-routing' })
+    text: t('nav.fellowship', 'Fellowship', { ns: 'apps-routing' }),
+    useCounter
   };
 }