+++ packages/page-accounts/src/modals/ProxyOverview.tsx
@@ -5,7 +5,7 @@ import type { ApiPromise } from '@polkadot/api';
 import type { SubmittableExtrinsic } from '@polkadot/api/types';
 import type { BatchOptions } from '@polkadot/react-hooks/types';
 import type { AccountId } from '@polkadot/types/interfaces';
-import type { NodeRuntimeProxyType, PalletProxyProxyDefinition } from '@polkadot/types/lookup';
+import type { KitchensinkRuntimeProxyType, PalletProxyProxyDefinition } from '@polkadot/types/lookup';
 import type { BN } from '@polkadot/util';
 
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
@@ -17,7 +17,7 @@ import { BN_ZERO } from '@polkadot/util';
 
 import { useTranslation } from '../translate';
 
-type PrevProxy = [AccountId | null, NodeRuntimeProxyType];
+type PrevProxy = [AccountId | null, KitchensinkRuntimeProxyType];
 
 interface Props {
   className?: string;
@@ -40,13 +40,13 @@ interface NewProxyProps extends ValueProps {
 }
 
 interface PrevProxyProps extends ValueProps {
-  onRemove: (accountId: AccountId, type: NodeRuntimeProxyType, index: number) => void;
+  onRemove: (accountId: AccountId, type: KitchensinkRuntimeProxyType, index: number) => void;
 }
 
 const BATCH_OPTS: BatchOptions = { type: 'all' };
 const EMPTY_EXISTING: [PalletProxyProxyDefinition[], BN] = [[], BN_ZERO];
 
-function createAddProxy (api: ApiPromise, account: AccountId, type: NodeRuntimeProxyType, delay = 0): SubmittableExtrinsic<'promise'> {
+function createAddProxy (api: ApiPromise, account: AccountId, type: KitchensinkRuntimeProxyType, delay = 0): SubmittableExtrinsic<'promise'> {
   return api.tx.proxy.addProxy.meta.args.length === 2
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore old version
@@ -54,7 +54,7 @@ function createAddProxy (api: ApiPromise, account: AccountId, type: NodeRuntimeP
     : api.tx.proxy.addProxy(account, type, delay);
 }
 
-function createRmProxy (api: ApiPromise, account: AccountId, type: NodeRuntimeProxyType, delay = 0): SubmittableExtrinsic<'promise'> {
+function createRmProxy (api: ApiPromise, account: AccountId, type: KitchensinkRuntimeProxyType, delay = 0): SubmittableExtrinsic<'promise'> {
   return api.tx.proxy.removeProxy.meta.args.length === 2
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore old version
@@ -150,12 +150,12 @@ function NewProxy ({ index, onChangeAccount, onChangeType, onRemove, proxiedAcco
   );
 }
 
-function getProxyTypeInstance (api: ApiPromise, index = 0): NodeRuntimeProxyType {
+function getProxyTypeInstance (api: ApiPromise, index = 0): KitchensinkRuntimeProxyType {
   // This is not perfect, but should work for `{Kusama, Node, Polkadot}RuntimeProxyType`
   const proxyNames = api.registry.lookup.names.filter((name) => name.endsWith('RuntimeProxyType'));
 
   // fallback to previous version (may be Substrate default), when not found
-  return api.createType<NodeRuntimeProxyType>(proxyNames.length ? proxyNames[0] : 'ProxyType', index);
+  return api.createType<KitchensinkRuntimeProxyType>(proxyNames.length ? proxyNames[0] : 'ProxyType', index);
 }
 
 function getProxyOptions (api: ApiPromise): { text: string; value: number; }[] {
@@ -187,7 +187,7 @@ function ProxyOverview ({ className, onClose, previousProxy: [existing] = EMPTY_
   useEffect((): void => {
     setBatchAdded(
       added
-        .filter((f): f is [AccountId, NodeRuntimeProxyType] => !!f[0])
+        .filter((f): f is [AccountId, KitchensinkRuntimeProxyType] => !!f[0])
         .map(([newAccount, newType]) => createAddProxy(api, newAccount, newType))
     );
   }, [api, added]);
@@ -216,7 +216,7 @@ function ProxyOverview ({ className, onClose, previousProxy: [existing] = EMPTY_
   );
 
   const _delPrev = useCallback(
-    (accountId: AccountId, type: NodeRuntimeProxyType, index: number): void => {
+    (accountId: AccountId, type: KitchensinkRuntimeProxyType, index: number): void => {
       setPrevious((previous) => previous.filter((_, i) => i !== index));
       setBatchPrevious((previous) => [...previous, createRmProxy(api, accountId, type)]);
     },