+++ packages/react-api/src/Api.tsx
@@ -13,7 +13,7 @@ import store from 'store';
 
 import { ApiPromise, WsProvider } from '@polkadot/api';
 import { deriveMapCache, setDeriveCache } from '@polkadot/api-derive/util';
-import { ethereumChains, typesBundle, typesChain } from '@polkadot/apps-config';
+import { ethereumChains, typesBundle } from '@polkadot/apps-config';
 import { web3Accounts, web3Enable } from '@polkadot/extension-dapp';
 import { TokenUnit } from '@polkadot/react-components/InputNumber';
 import { StatusContext } from '@polkadot/react-components/Status';
@@ -265,8 +265,7 @@ async function createApi (apiUrl: string, signer: ApiSigner, onError: (error: un
       registry,
       signer,
       types,
-      typesBundle,
-      typesChain
+      typesBundle
     });
 
     // See https://github.com/polkadot-js/api/pull/4672#issuecomment-1078843960